+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+## [0.1.3] - 2020-02-01
+- Support Polystar shape
+- Reorganize examples.
+
 ## [0.1.2] - 2020-01-31
-Implement `Lottie.network`, `Lottie.file` and `Lottie.memory`
+- Implement `Lottie.network`, `Lottie.file` and `Lottie.memory`
 
 ## [0.1.1] - 2020-01-31
-Fix analysis lints
+- Fix analysis lints
 
 ## [0.1.0] - 2020-01-31
-Initial conversion of [lottie-android](https://github.com/airbnb/lottie-android) to Dart/Flutter
+- Initial conversion of [lottie-android](https://github.com/airbnb/lottie-android) to Dart/Flutter